"
This mini-series contains a few trivial fixes to be able
to build scylla on Fedora 36 Pre-Release, which will
soon enter "Beta" state.

It's mostly fixes due to some changes to external dependencies,
e.g. boost.outcome and libfmt.

Tests: unit(dev)
"

* 'fc36_build_fixes_v1' of https://github.com/ManManson/scylla:
  schema: fix build issues with libstdc++ 12
  treewide: fix compilation issues with fmtlib 8.1.0+
  utils/result.hh: add missing header includes for boost.outcome

+++ alternator/server.cc
@@ -611,7 +611,7 @@ future<> server::json_parser::stop() {
 
 const char* api_error::what() const noexcept {
     if (_what_string.empty()) {
-        _what_string = format("{} {}: {}", _http_code, _type, _msg);
+        _what_string = format("{} {}: {}", static_cast<int>(_http_code), _type, _msg);
     }
     return _what_string.c_str();
 }
